+++ react_client/src/components/animation/Animation.jsx
@@ -1,11 +1,10 @@
-
 import ReactLoading from "react-loading";
 
 // Loading animation (GH)
 export const spinAnimation = function reactSpinLoadingAnimation() {
   return (
     <ReactLoading
-      className="mx-auto"
+      className="mx-auto my-3"
       type="spin"
       color="#4a4537"
       height={100}
